Feito verificação se o cliente tem vínculo com devedores antes de excluir.

+++ src/clientes/clientes.service.ts
@@ -1,20 +1,22 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { ClienteDevedor } from 'src/clientes-devedores/cliente-devedor.entity';
 import { Repository } from 'typeorm';
 import { Cliente } from './cliente.entity';
 
 @Injectable()
 export class ClientesService {
     constructor(
-        @InjectRepository(Cliente) private clientesRepository: Repository<Cliente>
+        @InjectRepository(Cliente) private clientesRepository: Repository<Cliente>,
+        @InjectRepository(ClienteDevedor) private clienteDevedorRepository: Repository<ClienteDevedor>
     ) {}
 
     async findAll(): Promise<Cliente[]> {
         return await this.clientesRepository.find();
     }
 
-    findOne(id: string): Promise<Cliente> {
-        return this.clientesRepository.findOne(id);
+    async findOne(id: string): Promise<Cliente> {
+        return await this.clientesRepository.findOne(id);
     }
 
     async create(cliente: Cliente): Promise<Cliente> {    
@@ -27,14 +29,45 @@ export class ClientesService {
         return await this.clientesRepository.findOne(id);
     }
 
-    async remove(id: string): Promise<boolean> {
-        const response = await this.clientesRepository.delete(id);
+    async remove(id: string): Promise<any> {
+        const hasDebit = await this.clientHasDebit(Number(id));
 
-        if(response) {
-            return true;
+        if(!hasDebit) {
+            const response = await this.clientesRepository.delete(id);
+
+            if(response) {
+                return {
+                    status: 'success',
+                    message: 'Cliente excluído com sucesso.'
+                }
+            }
+            else {
+                return {
+                    status: 'error',
+                    message: 'Problemas ao excluir cliente.'
+                }
+            }
         }
         else {
+            return {
+                status: 'error',
+                message: 'Não foi possível excluir, pois este cliente tem vínculo com débitos de devedores.'
+            }
+        }
+    }
+
+    async clientHasDebit(clientId: number): Promise<boolean> {
+        const hasDebit = await this.clienteDevedorRepository.find({
+            where: {
+                cliente: clientId
+            }
+        });
+
+        if(hasDebit.length == 0) {
             return false;
+        }
+        else {
+            return true;
         }
     }
 }
